+++ packages/leaflet-map-adapter/src/leaflet.d.ts
@@ -1,0 +1,6 @@
+declare namespace L {
+  export interface Map {
+    _controlContainer: HTMLElement;
+    _addUnselectCb: (args: () => void) => void;
+  }
+}