+++ packages/mapboxgl-map-adapter/src/util/makeHtmlFromString.ts
@@ -1,0 +1,5 @@
+export function makeHtmlFromString(str: string): HTMLElement {
+  const html = document.createElement('div');
+  html.innerHTML = str;
+  return html;
+}