+++ packages/ngw-kit/src/index.ts
@@ -47,6 +47,7 @@
  * ```
  *
  * @packageDocumentation
+ * @module ngw-kit
  */
 import { NgwKit } from './NgwKit';
 
@@ -57,16 +58,20 @@ export * from './IdentifyItem';
 
 export * from './utils/utils';
 export * from './utils/identifyUtils';
-export * from './utils/getCompanyLogo';
-export * from './utils/featureLayerUtils';
-export * from './utils/fetchNgwLayerFeature';
-export * from './utils/fetchNgwLayerFeatures';
-export * from './utils/fetchNgwLayerFeatureCollection';
-export * from './utils/fetchNgwLayerItem';
-export * from './utils/fetchNgwLayerItems';
 export * from './utils/fetchNgwExtent';
-export * from './utils/resourceIdFromLayerOptions';
+export * from './utils/getCompanyLogo';
 export * from './utils/parseDateFromNgw';
+export * from './utils/featureLayerUtils';
+export * from './utils/fetchNgwLayerItem';
+export * from './utils/createPopupContent';
+export * from './utils/fetchNgwLayerItems';
+export * from './utils/fetchNgwLayerFeature';
+export * from './utils/uploadFeatureAttachment';
+export * from './utils/mapFeatureDisplayName';
+export * from './utils/fetchNgwLayerFeatures';
+export * from './utils/resourceIdFromLayerOptions';
+export * from './utils/fetchNgwLayerFeatureCollection';
+export * from './utils/prepareNgwFieldsToPropertiesFilter';
 
 export * from './adapters/createOnFirstShowNgwAdapter';
 export * from './adapters/createOnFirstShowAdapter';