+++ packages/ol-map-adapter/src/layer-adapters/ImageAdapter.ts
@@ -17,7 +17,7 @@ export class ImageAdapter implements MainLayerAdapter {
 
   constructor(public map: Map, public options: ImageAdapterOptions) {}
 
-  addLayer(options: ImageAdapterOptions): ImageLayer | undefined {
+  addLayer(options: ImageAdapterOptions): ImageLayer<ImageWMS> | undefined {
     if (options.url) {
       const ratio = options.ratio !== undefined ? options.ratio : 1;
       const imageOptions: ImageWMSOptions = {
@@ -32,9 +32,8 @@ export class ImageAdapter implements MainLayerAdapter {
       imageOptions.imageLoadFunction = (image, src) => {
         const url = src.split('?')[0];
         const query = src.split('?')[1];
-        const { resource, BBOX, WIDTH, HEIGHT, ...params } = queryToObject(
-          query,
-        );
+        const { resource, BBOX, WIDTH, HEIGHT, ...params } =
+          queryToObject(query);
 
         const queryParams = {
           resource,
@@ -75,12 +74,13 @@ export class ImageAdapter implements MainLayerAdapter {
             return super.renderFrame(frameState, target);
           }
         }
-        class ImageLayerExtended extends ImageLayer {
+        class ImageLayerExtended extends ImageLayer<ImageWMS> {
           // @ts-ignore
           createRenderer() {
             return new CanvasILRendererExtended(this);
           }
         }
+        // @ts-ignore
         ImLayer = ImageLayerExtended;
 
         class ImageWmsExtended extends ImageWMS {