+++ packages/webmap/src/interfaces/MapState.ts
@@ -1,8 +1,6 @@
 import type { WebMapEvents } from './Events';
 
-/**
- * @public
- */
+
 export interface MapStateItem<V extends any = any> {
   event: keyof WebMapEvents;
   name: string;
@@ -11,7 +9,5 @@ export interface MapStateItem<V extends any = any> {
   parse: (str: string) => V;
 }
 
-/**
- * @public
- */
+
 export type MapState = MapStateItem[];